requirements.txt
djangorestframework==3.16.1
Django==6.0.1
django_cors_headers==4.9.0
requests==2.32.5